refactor: extract comments logic to custom hook
