+++ client/src/app/postview/[id]/page.tsx
@@ -3,30 +3,28 @@ import { useParams } from "next/navigation"
 import { useState, useEffect } from "react";
 import Header from "../../components/layout/header";
 import { Post } from "@/types/posts";
-import { Comment } from "@/types/comments";
-import { FormEvent } from "react";
 import { fetchPostById } from "../fetchpostid";
 import EditButton from "@/app/posts/editbutton";
 import CommentForm from "../commentform";
-import { getApiUrl } from '../../services/api';
-import useAuth from "../../hooks/useauth";
+import useAuth from "../../hooks/useAuth";
 import CommentsList from "../commentlist";
+import useComments from "../../hooks/useComments";
 
 function PostView() {
-    const { id } = useParams()
-    const postId = id as string;
+    const { id: postId } = useParams() as { id: string };
     const [post, setPost] = useState<Post | null>(null);
     const [loading, setLoading] = useState(true);
     const [error, setError] = useState<string | null>(null);
     const { user, token } = useAuth();
 
-    const [comments, setComments] = useState<Comment[]>([]);
-    const [commentTitle, setCommentTitle] = useState("");
-    const [commentContent, setCommentContent] = useState("");
-    const [commentLoading, setCommentLoading] = useState(true);
-    const [commentError, setCommentError] = useState<string | null>(null);
-    
-    const [message, setMessage] = useState<string | null>(null);
+    const {
+    comments,
+    commentTitle, setCommentTitle,
+    commentContent, setCommentContent,
+    postComment, deleteComment,
+    commentLoading, commentError, message,
+    } = useComments(postId, user?.userid || "", token, setError);
+
     const [showForm, setShowForm] = useState(false);
     const [activeCommentId, setActiveCommentId] = useState<number | null>(null);
 
@@ -45,80 +43,6 @@ function PostView() {
 
         loadPost();
     }, [postId]);
-
-
-    useEffect(() => {
-        const fetchComments = async () => {
-            try {
-                const data = await fetch(getApiUrl(`/comments/post/${postId}`));
-                const comments = await data.json();
-                setComments(comments);
-                setCommentLoading(false);
-            } catch (error) {
-                console.error("Error fetching comments:", error);
-                setCommentError("Failed to fetch comments");
-                setCommentLoading(false);
-            }
-        }
-        fetchComments();
-    }, [postId]);
-
-    async function postComment(event: FormEvent<HTMLFormElement>) {
-        setError("")
-        setMessage("")
-        setCommentError("")
-        event.preventDefault();
-        const json = {
-            title: commentTitle,
-            content: commentContent,
-            postId: Number(postId),
-            userId: user?.userid || "",
-        };
-        try {
-            const res = await fetch(getApiUrl(`/comments/post/${postId}`), {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${token}`,
-                },
-                body: JSON.stringify(json),
-            });
-            const data = await res.json();
-            if (!res.ok) {
-                setCommentError(data.message || "Failed to post comment");
-                return;
-            }
-            setComments((prevComments) => [...prevComments, data]);
-            setMessage("Comment posted successfully!");
-            setCommentLoading(false);
-        } catch (error) {
-            console.error("Error posting comment:", error);
-            setCommentError("Failed to post comment");
-        }
-    }
-
-    async function deleteComment(id: number) {
-        try {
-            const res = await fetch(getApiUrl(`/comments/${id}`), {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${token}`,
-                },
-            });
-            const data = await res.json();
-            if (!res.ok) {
-                setCommentError(data.message || "Failed to delete comment");
-                return;
-            }
-            setComments((prevComments) => prevComments.filter((comment) => comment.id !== id));
-            setMessage("Comment deleted successfully!");
-        } catch (error) {
-            console.error("Error deleting comment:", error);
-            setCommentError("Failed to delete comment");
-        }
-    }
-
 
     if (loading) {
         return <div>Loading...</div>;
